requirements.txt
pandas>=2.1
numpy>=1.26
networkx>=3.2
matplotlib>=3.8
plotly>=5.23
tqdm>=4.66
# Optional (comment out if you do not need maps)
geopandas>=0.14
folium>=0.15
pyproj>=3.6
# Streamlit app
streamlit>=1.38
pyyaml>=6.0
